requirements.txt
Django==2.1.7
graphene-django==2.2.0
postgres==2.2.2
psycopg2-binary==2.7.7
pytz==2018.9
flake8==3.5.0
PyJWT==1.7.1
djoser==1.1.5
pytest==4.4.0
cloudinary==1.15.0
django-enumfields==1.0.0
Pillow==6.0.0
pytest-cov==2.6.1
pytest-django==3.1.2
django-graphql-extensions==0.0.6
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
